+++ packages/common/src/index.ts
@@ -1,11 +1,17 @@
 export * from "./common";
 export * from "./createBurnerAccount";
-export * from "./createContract";
 export * from "./createNonceManager";
+export * from "./getContract";
 export * from "./getBurnerPrivateKey";
+export * from "./getNonceManager";
+export * from "./getNonceManagerId";
 export * from "./hexToResourceId";
 export * from "./readHex";
 export * from "./resourceIdToHex";
 export * from "./resourceTypes";
 export * from "./spliceHex";
 export * from "./transportObserver";
+export * from "./writeContract";
+
+/** @deprecated use `getContract` instead */
+export { createContract } from "./createContract";
